+++ extensions/git/src/components/actions/WorkspaceNavigationActions.tsx
@@ -3,6 +3,7 @@ import { GitView } from "../../types";
 import { Action, ActionPanel, Icon, List } from "@raycast/api";
 import { RepositoryDirectoryActions } from "./RepositoryDirectoryActions";
 import { SwitchRepositoryAction } from "./SwitchRepositoryAction";
+import { GitConfigView } from "../views/GitConfigView";
 
 export function WorkspaceNavigationDropdown(context: NavigationContext) {
   return (
@@ -21,12 +22,7 @@ export function WorkspaceNavigationDropdown(context: NavigationContext) {
       <List.Dropdown.Item title="Branches" value="branches" keywords={["graph", "remote"]} icon={`git-branch.svg`} />
       <List.Dropdown.Item title="Tags" value="tags" icon={Icon.Tag} />
       <List.Dropdown.Item title="Remotes" value="remotes" keywords={["origin"]} icon={Icon.Network} />
-      <List.Dropdown.Item
-        title="Submodules"
-        value="submodules"
-        keywords={["submodule"]}
-        icon={`submodule-folder.svg`}
-      />
+      <List.Dropdown.Item title="Submodules" value="submodules" icon={`submodule-folder.svg`} />
       <List.Dropdown.Item title="Stashes" value="stashes" keywords={["bookmark"]} icon={Icon.Bookmark} />
       <List.Dropdown.Item
         title="Files"
@@ -93,6 +89,14 @@ export function WorkspaceNavigationActions(context: NavigationContext & Reposito
       </ActionPanel.Section>
 
       <RepositoryDirectoryActions repositoryPath={context.gitManager.repoPath} />
+
+      <Action.Push
+        title="Show Git Config"
+        target={<GitConfigView {...context} />}
+        icon={Icon.Gear}
+        shortcut={{ modifiers: ["cmd", "shift"], key: "," }}
+      />
+
       <SwitchRepositoryAction repositoryPath={context.gitManager.repoPath} onSwitch={context.switchTo} />
     </>
   );